Avoid creating or amending a commit if not necessary

+++ scripts/pull-and-resolve-english-changes.sh
@@ -3,9 +3,10 @@ set -euo pipefail
 
 function fail { >&2 echo -e "ERROR: $1"; exit 1; }
 
-(( $# == 1 )) || fail "Wrong number of arguments\nUsage: $0 <solidity repo commit/branch/tag>"
+(( $# == 2 )) || fail "Wrong number of arguments\nUsage: $0 <solidity repo commit/branch/tag> <commit message>"
 git config remote.english.url || fail "Remote 'english' is not defined"
 SOLIDITY_REF="$1"
+COMMIT_MESSAGE="$2"
 
 # ASSUMPTIONS:
 # - We're in the root directory of a working copy of the translation repository.
@@ -34,25 +35,28 @@ echo "branch_exists=$branch_exists" >> "$GITHUB_OUTPUT"
 # We want to include the conflict markers as a part of the merge commit so that they're easy to spot in the PR.
 # The command will also fail if in the main repo there were modifications to files outside
 # of docs/ (these files are deleted in translation repos). These are the conflicts we want to ignore.
-git merge "english/${SOLIDITY_REF}" || true
+if ! git merge "english/${SOLIDITY_REF}" -m "$COMMIT_MESSAGE"; then
+    # Unstage everything without aborting the merge.
+    # This also "resolves" conflicts by keeping conflicted files as is including the conflict markers.
+    git reset .
 
-# Unstage everything without aborting the merge.
-# This also "resolves" conflicts by keeping conflicted files as is including the conflict markers.
-git reset .
+    # The only changes from the main repo that we're interested in are those in docs/ and
+    # CMakeLists.txt, which is parsed by our Sphinx config to determine version. Stage them.
+    git add docs/ CMakeLists.txt
 
-# The only changes from the main repo that we're interested in are those in docs/ and
-# CMakeLists.txt, which is parsed by our Sphinx config to determine version. Stage them.
-git add docs/ CMakeLists.txt
+    # Reset any files seen by git as modified/deleted to the state from before the merge.
+    # We need this for files outside of docs/ that happen to match the path of some file that exists
+    # in the main repo, for example README.md or .gitignore. We do not copy these when setting up a
+    # translation repo but to git they look like files from the main repo that need to be synced.
+    files_to_reset=$(git ls-files --modified --deleted)
+    if [[ $files_to_reset != "" ]]; then
+        # NOTE: If the file list passed to `git checkout` is empty, git aborts the merge. We don't want that.
+        git ls-files --modified --deleted | xargs git checkout --
+    fi
 
-# Reset any files seen by git as modified/deleted to the state from before the merge.
-# We need this for files outside of docs/ that happen to match the path of some file that exists
-# in the main repo, for example README.md or .gitignore. We do not copy these when setting up a
-# translation repo but to git they look like files from the main repo that need to be synced.
-files_to_reset=$(git ls-files --modified --deleted)
-if [[ $files_to_reset != "" ]]; then
-    # NOTE: If the file list passed to `git checkout` is empty, git aborts the merge. We don't want that.
-    git ls-files --modified --deleted | xargs git checkout --
+    # All the other files from the original merge commit are now untracked. We don't want them in the PR.
+    git clean -d --force
+
+    # Finalize the merge. Should create a commit with the message we specified in the original command.
+    git commit --no-edit
 fi
-
-# All the other files from the original merge commit are now untracked. We don't want them in the PR.
-git clean -d --force
